Adding a script to deal with "EOLE" based servers.
This script simply creates a eole configuration file called config.eol
with informations provided by contextualisation.

It use the IPs setted by 00-Network and it allows users to fill EOLE variables
using contextualisation variables.

If a variable is prefixed by CR_ the script creates an entry in config.eol
with the given value. If this value match this format : {value0,value1,valueX}
the script will fill the variable with the list in the good format.

This way if you use an EOLE server you can fill the configuration from the ONE context.