B #118: one-context-force breaks boot status
We must make sure "one-context-force.service" is not run during boot
disk detection but on hotplug events.

When booting with a swap partition on the disk, the udev rules will
start the "one-context-force" service long before "one-context-local"
and "one-context".

This results in the system being in "degraded" state as reported by
"systemctl is-system-running".

* src/usr/lib/systemd/system/one-context-force.service##systemd.one:
  add conditions to run only on hotplug events.

Signed-off-by: Daniel Dehennin <daniel.dehennin@ac-dijon.fr>