M #-: Drop VMware poweroff script on Alpine
This file was migrated from dedicated Alpine contextualization
repository. The reason why it was necessary to do explicit poweroff on
VMware poweroff action is unknown. Currently, it happens that VM with
this file doesn't properly poweroff, but just halts. So it looks like a better
option to drop this legacy file and depend on built-in logic.