F #265: Improve START_SCRIPT
Scripts run by one-contexd will not be executed in a subshell - this
will fix the issue with the START_SCRIPT which would hold the boot if
any long-running process would originate in it (e.g. daemon/service).

User still will need to do something like this inside the START_SCRIPT
(or any other context script) otherwise SIGTERM will be sent:

    nohup service nginx start &

Signed-off-by: Petr Ospalý <pospaly@opennebula.io>